+++ src/common/nodes/ComponentInterface.ts
@@ -108,4 +108,16 @@ export class ComponentInterface extends NamedNode<ComponentInterface> implements
         return componentInterface;
     }
 
+    /**
+     * marks this node as deleted
+     * this also marks this node as changed
+     */
+    public async markDeleted(): Promise<void> {
+        if(!this.isDeleted) {
+            await super.markDeleted();
+            await this.consumedByProperty.clear();
+            await (await this.component()).interfacesProperty.remove(this);
+        }
+    }
+
 }
